requirements.txt
asgiref==3.8.1
Django==4.2.10
django-environ==0.11.2
djangorestframework==3.15.1
sqlparse==0.5.3
tzdata==2025.2
psycopg==3.2.6
psycopg2==2.9.10
celery==5.4.0
redis==5.0.4
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1 
django-anymail==10.1
dj-rest-auth==2.2.6
django-allauth==0.49.0
social-auth-app-django==5.4.0

